requirements.txt
fastai==2.8.1              
torch==2.6.0              
torchvision==0.18.0       
pandas==2.2.2             
matplotlib==3.9.0         
seaborn==0.14.1           
scikit-learn==1.5.0       
ipykernel==6.29.4         
jupyter==1.0.0            
fastapi==0.111.0          
uvicorn==0.29.0           
python-multipart==0.0.9   
pillow==10.3.0            
python-dotenv==1.0.1      
cloudpickle==3.0.0       